advertis_service/requirements.txt
# FastAPI server and utilities
fastapi
uvicorn[standard]
python-dotenv

# Database and Cache clients
redis

# Core AI and Vector Store libraries
langchain
langgraph
langchain-openai
chromadb